feat: set up file autoloader for functions as well
