+++ src/Bojler/game_status.php
@@ -11,7 +11,6 @@ use Ds\Set;
 
 
 require_once __DIR__ . '/db.php'; # TODO DatabaseHandler, DictionaryType with PSR-4 autoloader
-require_once __DIR__ . '/util.php'; # TODO remove_special_char with PSR-4 autoloader
 require_once __DIR__ . '/player.php'; # TODO PlayerHandler with PSR-4 autoloader
 
 # TODO better dependency injection surely...
