+++ src/Bojler/player.php
@@ -7,43 +7,6 @@ use Discord\Parts\{
     User\Member,
 };
 
-function hungarian_role(Member $member)
-{
-    $hungarian_roles = ['Beginner', 'Native speaker', 'Intermediate', 'Fluent', 'Advanced', 'Distant native'];
-    $roles = $member->roles;
-    foreach ($roles as $item) {
-        foreach ($hungarian_roles as $hu_role) {
-            if ($item->name === $hu_role) {
-                return $hu_role;
-            }
-        }
-    }
-    return '';
-}
-
-function name_shortened($name)
-{
-    if (grapheme_strlen($name) >= 15) {
-        $name = grapheme_substr($name, 0, grapheme_strpos($name, '|') ?: null);
-    }
-    if (grapheme_strlen($name) >= 15) {
-        $name = grapheme_substr($name, 0, grapheme_strpos($name, ' ') ?: null);
-    }
-    if (grapheme_strlen($name) >= 15) {
-        $name = grapheme_substr($name, 0, 15);
-    }
-    return $name;
-}
-
-function discord_specific_fields(Member $member)
-{
-    return [
-        'name' => $member->username,
-        'role' => hungarian_role($member),
-        'server_name' => name_shortened($member->nick ?? $member->username)
-    ];
-}
-
 class PlayerHandler
 {
     private static $instance;
